Updated validation of arguments due to redundancies or new arguments.
arg_handling.rs
Updated Args struct to make the input parameter required.
Updated Args struct to make the metadata parameterconflict with extract.
Added logic under extraction mode for checking if the output path is a directory and throw an error that it must be a directory  if met.
Added logic under extraction mode for checking if the force flag is used and throw an error that it can't be used under extraction mode.
Added logic for checking for the existance of input paths. If a path does not exist, throw an error.

storage_io.rs:
Remove dir_contains_files function as it is no longer needed.