Updated readme and added file specification.
README.md:
Added .ssmc file format section and where to file the full specification.
Corrected some formatting in the Command-Line Arguments table.
Updated the license section on where to find licenses for this project and that the file specification is under a different license.

Added folder containing license and file specification.