Refactor archive finalization to a builder pattern
Library:
archive.rs:
Made ArchiveBuilder struct to be used by the ArchiveBuilder
Refactored the finalize_archive function to use a builder pattern and use optional paramters to be easier to use for devs.
Update function documentation to be in line with ArchiveBuilder and provided function documentation for each helper function for the optional parameters.

lib.rs:
Update/replace public use of finalize_archive to be ArchiveBuilder.

cli_application:
compress.rs:
Replace finalize_archive in use of sprite_shrink library to be ArchiveBuilder.
Refactor use of finalize_archive to use new ArchiveBuilder.