Reworked documentation and ordering of ffi_structs.rs
library:
archive.rs:
Changed CUserData to FFIUserData due to change in ffi_structs to be consistent with FFI struct naming.

ffi_structs.rs:
Added documentation to the top of the file for providing insight into purpose of file.
Reordered all structs to be in alphabetical order.
Updated or added documentation for each struct.