Implemented error enum for serialization_ffi.rs
library:
serialization_ffi.rs:
Modified the signature and return type for serialize_uncompressed_data_ffi to return an error code via the error enum and reworked logic to return the pointer and FFIStatus.
Added check for input ptrs and if they are null to fail early.

mod.rs:
Added additional error to the FFI status enum due to the additional errors handled in serialization_ffi.rs.